Modify the packet sending process to transmit the size as a ushort type instead of an int type

+++ Assets/2.Script/Network/ServerSession.cs
@@ -60,7 +60,7 @@ public class ServerSession
 
         byte[] sendBuffer = new byte[size + 4];
 
-        Array.Copy(BitConverter.GetBytes(size + 4), 0, sendBuffer, 0, sizeof(ushort));
+        Array.Copy(BitConverter.GetBytes((ushort)(size + 4)), 0, sendBuffer, 0, sizeof(ushort));
         Array.Copy(BitConverter.GetBytes((ushort)packetID), 0, sendBuffer, 2, sizeof(ushort));
         Array.Copy(packet.ToByteArray(), 0, sendBuffer, 4, size);
